separate demo js
